% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Local entropy computation}
\usage{
entropy(X, alpha = 2, K, scale = F)
}
\arguments{
\item{X}{is the input matrix of n x d. n are the observations (e.g.:
projected curves) and d is the dimension of X (dimension of the projection
space).}

\item{alpha}{a parameter defining the entropy function. Default = 2.}

\item{K}{number of neighbour points to consider in the computation of
entropy.}

\item{scale}{logical variable indicating if scaling is required. Default =
FALSE.}
}
\value{
\item{local.entropy}{local entropy relative to each point in the
sample.} \item{entropy}{estimated entropy.}
}
\description{
This function allows you to compute the family of local alpha entropy as
stated in Martos et al (2018).
}
\details{
The function computes the alpha entropy and the local alpha entropy (see
reference for further details) of a data set using a non parametric density
estimator.
}
\references{
Martos, G. et al (2018): Entropy Measures for Stochastic
Processes with Applications in Functional Anomaly Detection. Entropy 20(1):
33 (2018).
}
