% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata_rkhs.R
\name{fdata_rkhs}
\alias{fdata_rkhs}
\title{functional data in rkhs}
\usage{
fdata_rkhs(curves, rk, gamma = 1e-05)
}
\arguments{
\item{curves}{a data matrix with observations (curves) in rows and the
discretizations points in columns.}

\item{rk}{kernek function rk object.}

\item{gamma}{regularization parameter. Defaoult value = 1e-5.}
}
\value{
\item{data}{input curves.} \item{fdata}{smoothed curves.}
\item{lambda}{coefficients of the (stable) and d dimensional RKHS
representation.} \item{alpha}{coefficients of the RKHS expansion.}
\item{gamma}{regularization parameter.}
}
\description{
Representing functinal data using Reproducing Kernel Hilbert Spaces.
Approximate each curve with a smooth function using a kernel function.
}
\details{
With this function each function can be represented with a vector in R^d.
}
\examples{
require(ftsa)  
curves = t(sqrt(ftsa::pm_10_GR$y))
t = 1:ncol(curves)
kernel = fpcb::rk(grid = t,sigma = 0.01)
f.data <- fpcb::fdata_rkhs(curves, rk = kernel)
matplot(t,t(curves), xlab='time', ylab='PM10 dataset', col='gray', lty=1, type='l')
matlines(t,t(f.data$fdata), col='blue', lty=1)
legend("topright",lty=c(1,1), col=c('gray','blue'), legend=c('PM10 curves','PM10 smoothed curves'))
}
\references{
A. Muñoz, J. González, Representing functional data using
support vector machines, Pattern Recognition Letters 31 (2010) 511–516. <doi:10.1016/j.patrec.2009.07.014>.
}
\author{
N. Hernández and J. Cugliari
}
