% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpcb.-package.R
\docType{package}
\name{fpcb-package}
\alias{fpcb-package}
\alias{fpcb}
\title{Predictive confindence bands for functional time series forecasting}
\description{
Functions to represent functional objects under a Reproducing Kernel Hilbert
Space (RKHS) framework as described in Muñoz & González (2010).
Autoregressive Hilbertian Model for functional time series using RKHS and
predictive confidence bands construction as proposed in Hernández et al
(2021).
}
\references{
\itemize{ \item A. Muñoz, J. González, Representing functional
data using support vector machines, Pattern Recog-nition Letters 31 (2010)
511–516.  \item Martos, G. et al (2018): Entropy Measures for Stochastic
Processes with Applications in Functional Anomaly Detection. Entropy 20(1):
33 (2018).  \item D. Wang, Z. Zhao, R. Willett, C. Y. Yau, Functional
autoregressive processes in reproducing kernel hilbert spaces, arXiv
preprint arXiv:2011.13993 (2020).  \item N. Hernández, J. Cugliari, J.
Jacques. Simultaneous Predictive Bands for Functional Time Series using
Minimum Entropy Sets. 2021. -Submited-. }
}
\author{
Nicolás Hernández [aut, cre], Jairo Cugliari [aut, cre], Julien Jacques [aut]\cr

Mantainer: Nicolás Hernández <nicolas.hernandez@mrc-bsu.cam.ac.uk>
}
\keyword{ARH}
\keyword{Bands}
\keyword{Confidence}
\keyword{Entropy}
\keyword{FTS}
\keyword{RKHS}
